move file to correct folder
