+++ app/react/V2/Routes/Settings/IX/components/PropertySidepanel.tsx
@@ -17,7 +17,7 @@ import {
   handleEntitySave,
   loadSidepanelData,
   SELECT_TYPES,
-} from './sidepanelFunctions';
+} from '../helpers';
 import { SidepanelForms } from './SidepanelForms';
 import { highlightsAtom, selectionErrorAtom, textSelectionAtom } from './atoms';
 import { selectAndSearchAtom } from './atoms/selectAndSearchAtom';
