+++ app/react/V2/Routes/Settings/IX/helpers/index.ts
@@ -5,3 +5,13 @@ export {
   formatAccepted,
 } from './helpers';
 export { updateSortingUrl } from './updateSortingUrl';
+export {
+  SELECT_TYPES,
+  coerceValue,
+  getFormValue,
+  loadSidepanelData,
+  loadValuesAndSuggestions,
+  handleEntitySave,
+  getPropertyNameFromExtractPair,
+  getTemplateFromExtractPair,
+} from './sidepanelFunctions';
